Fix compiler options: allow global scalac options, and got it to build again.
