+++ scalatest/src/main/scala/org/scalatest/matchers/should/Matchers.scala
@@ -7127,7 +7127,7 @@ org.scalatest.exceptions.TestFailedException: org.scalatest.Matchers$ResultOfCol
     // SKIP-DOTTY-START
     def shouldBe(aType: ResultOfATypeInvocation[_]): Assertion = macro TypeMatcherMacro.shouldBeATypeImpl
     // SKIP-DOTTY-END
-    //DOTTY-ONLY inline def shouldBe(aType: ResultOfATypeInvocation[_]): Assertion = ${ org.scalatest.matchers.should.TypeMatcherMacro.shouldBeATypeImpl('{this}, '{aType}) }
+    //DOTTY-ONLY inline def shouldBe(aType: ResultOfATypeInvocation[_]): Assertion = ${ TypeMatcherMacro.shouldBeATypeImpl('{this}, '{aType}) }
 
     /**
      * This method enables syntax such as the following:
@@ -7140,7 +7140,7 @@ org.scalatest.exceptions.TestFailedException: org.scalatest.Matchers$ResultOfCol
     // SKIP-DOTTY-START
     def shouldBe(anType: ResultOfAnTypeInvocation[_]): Assertion = macro TypeMatcherMacro.shouldBeAnTypeImpl
     // SKIP-DOTTY-END
-    //DOTTY-ONLY inline def shouldBe(anType: ResultOfAnTypeInvocation[_]): Assertion = ${ org.scalatest.matchers.should.TypeMatcherMacro.shouldBeAnTypeImpl('{this}, '{anType}) }
+    //DOTTY-ONLY inline def shouldBe(anType: ResultOfAnTypeInvocation[_]): Assertion = ${ TypeMatcherMacro.shouldBeAnTypeImpl('{this}, '{anType}) }
 
     /**
      * This method enables syntax such as the following:
@@ -7615,7 +7615,7 @@ org.scalatest.exceptions.TestFailedException: org.scalatest.Matchers$ResultOfCol
     // SKIP-DOTTY-START
     def should(compileWord: CompileWord)(implicit pos: source.Position): Assertion = macro CompileMacro.shouldCompileImpl
     // SKIP-DOTTY-END
-    // FIXME //DOTTY-ONLY inline def should(compileWord: CompileWord)(implicit pos: source.Position): Assertion = ${ CompileMacro.shouldCompileImpl('{this}, '{compileWord})('{pos}) }
+    //DOTTY-ONLY inline def should(compileWord: CompileWord)(implicit pos: source.Position): Assertion = ${ CompileMacro.shouldCompileImpl('{this}, '{compileWord})('{pos}) }
 
     /**
      * This method enables syntax such as the following:
@@ -7628,7 +7628,7 @@ org.scalatest.exceptions.TestFailedException: org.scalatest.Matchers$ResultOfCol
     // SKIP-DOTTY-START
     def shouldNot(compileWord: CompileWord)(implicit pos: source.Position): Assertion = macro CompileMacro.shouldNotCompileImpl
     // SKIP-DOTTY-END
-    // FIXME //DOTTY-ONLY inline def shouldNot(compileWord: CompileWord)(implicit pos: source.Position): Assertion = ${ CompileMacro.shouldNotCompileImpl('{this}, '{compileWord})('{pos}) }
+    //DOTTY-ONLY inline def shouldNot(compileWord: CompileWord)(implicit pos: source.Position): Assertion = ${ CompileMacro.shouldNotCompileImpl('{this}, '{compileWord})('{pos}) }
 
     /**
      * This method enables syntax such as the following:
@@ -7641,7 +7641,7 @@ org.scalatest.exceptions.TestFailedException: org.scalatest.Matchers$ResultOfCol
     // SKIP-DOTTY-START
     def shouldNot(typeCheckWord: TypeCheckWord)(implicit pos: source.Position): Assertion = macro CompileMacro.shouldNotTypeCheckImpl
     // SKIP-DOTTY-END
-    // FIXME //DOTTY-ONLY inline def shouldNot(typeCheckWord: TypeCheckWord)(implicit pos: source.Position): Assertion = ${ CompileMacro.shouldNotTypeCheckImpl('{this}, '{typeCheckWord})('{pos}) }
+    //DOTTY-ONLY inline def shouldNot(typeCheckWord: TypeCheckWord)(implicit pos: source.Position): Assertion = ${ CompileMacro.shouldNotTypeCheckImpl('{this}, '{typeCheckWord})('{pos}) }
 
 /*
     /**
